Add second section to index.html
This section goes just bellow the header section.

It features in the left side a "sub header", a description and a
button; on the right side goes a div that acts like a placeholder for
an image.

This section is prone to changes in the html elements it has in order
to have the desired presentation.