Add style to second and third sections
In the second section it is added some size, color and bold to the
header text; for the third section, it is added some background color,
font size to the text, color and italic to give the presentation of
a quote.

Also some rules have been grouped with certain elements to reduce the
number of lines of the CSS file.