Give more style to header and first section
In previous commits, first section only had changes related to colors,
fonts and alignment; now there are more modifications to the alignment
of the elements under the header.

Some of those changes are applying width to the divs holding all the
text and the one with the placeholder for an image.