Give style to second-section
With the help of the divs added before, this section now has its items
centered horizontally in the screen and the text goes just bellow each
div / image and that text is centered.

The new divs that replace the empty images now has a rounded blue
border to illustrate them on the page.

Also it is changed the vertical padding to the first section to 50px,
just to give less vertical space between all the sections of the page.